% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uPlot.R
\name{uplot}
\alias{uplot}
\title{Test}
\usage{
uplot(df_data, opts, width = NULL, height = NULL, elementId = NULL, digits = 4)
}
\arguments{
\item{df_data}{data.frame with data. First column is deciding time and needs to be ascending}

\item{opts}{list of options}

\item{width}{integer, widget width}

\item{height}{integer, widget height}

\item{elementId}{HTML container element ID}

\item{digits}{how many decimanl digits should JSON parser use? Use I() to specify significant digits. Use NA for max precision. Higher number results in much slower parsing. Default is 4.}
}
\description{
Test
}
\examples{

\dontrun{
df_data <- data.frame(x = 1:100,
                      y = rnorm(100))
options <- list(title = "My Chart",
                id = "chart1",
                class = "my-chart",
                width = 800,
                height = 600,
                scales = list(x = list(time = FALSE)),
                series = list(NULL,
                              list(stroke = "red",
                                   width = 1,
                                   label = "Test")))
uplot(df_data,
      options)
}
}
